+++ pkg/commands/integration/describe.go
@@ -1,0 +1,60 @@
+package integration
+
+import (
+	"context"
+	"io"
+
+	"github.com/fastly/go-fastly/v17/fastly"
+
+	"github.com/fastly/cli/pkg/argparser"
+	fsterr "github.com/fastly/cli/pkg/errors"
+	"github.com/fastly/cli/pkg/global"
+	"github.com/fastly/cli/pkg/text"
+)
+
+// DescribeCommand calls the Fastly API to describe a notification integration.
+type DescribeCommand struct {
+	argparser.Base
+	argparser.JSONOutput
+
+	Input fastly.GetIntegrationInput
+}
+
+// NewDescribeCommand returns a usable command registered under the parent.
+func NewDescribeCommand(parent argparser.Registerer, g *global.Data) *DescribeCommand {
+	c := DescribeCommand{
+		Base: argparser.Base{
+			Globals: g,
+		},
+	}
+	c.CmdClause = parent.Command("describe", "Retrieve a single notification integration").Alias("get")
+
+	// Required.
+	c.CmdClause.Arg("id", "Integration ID").Required().StringVar(&c.Input.ID)
+
+	// Optional.
+	c.RegisterFlagBool(c.JSONFlag())
+
+	return &c
+}
+
+// Exec invokes the application logic for the command.
+func (c *DescribeCommand) Exec(_ io.Reader, out io.Writer) error {
+	if c.Globals.Verbose() && c.JSONOutput.Enabled {
+		return fsterr.ErrInvalidVerboseJSONCombo
+	}
+
+	o, err := c.Globals.APIClient.GetIntegration(context.TODO(), &c.Input)
+	if err != nil {
+		c.Globals.ErrLog.Add(err)
+		return err
+	}
+
+	if ok, err := c.WriteJSON(out, o); ok {
+		return err
+	}
+
+	text.PrintIntegration(out, o)
+
+	return nil
+}